requirements.txt
celery==5.2.6
Django==3.2.7
drf-spectacular==0.22.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-filter==21.1
django-jazzmin==2.4.8
django-libsass==0.9
django-redis==5.2.0
requests==2.25.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-gettext==4.0.0
xmltodict==0.13.0
protobuf==3.17.0
google-api-core==1.28.0
google-api-python-client==2.7.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.7.2
google-cloud-translate==2.0.1
google-pasta==0.2.0
google-trans-new==1.1.9
googleapis-common-protos==1.53.0
django-modeltranslation==0.17.5
eventlet==0.29.0
django-phonenumber-field==6.3.0
phonenumbers==8.12.52
django-cors-headers==4.3.0
django-celery-beat==2.6.0
beautifulsoup4==4.12.3
lxml==5.3.0
gunicorn==21.2.0
